fixed overflow in number of remaining bytes
Since 2-byte append status is used, there is a chance the RX FIFO THR
interrupt is served after packet ending interrupt. For example it
happens when RX FIFO THR is set to 4 bytes and the number of remaining
bytes is 2.
In that case, number of bytes in RX FIFO must be checked before further
processes.